apps/inventory-manager/requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.15.0
discogs-client==2.3.0
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.4
Pillow==10.1.0
python-barcode==0.14.0 